core: reorganize the standalone sim endpoint
