+++ core/src/main/java/fr/sncf/osrd/standalone_sim/result/ResultStops.java
@@ -1,0 +1,7 @@
+package fr.sncf.osrd.standalone_sim.result;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+
+@SuppressFBWarnings("URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD")
+public record ResultStops(double time, double position, double duration) {
+}
